requirements.txt
jupyter-core==4.10.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
pandas==1.4.2
pickleshare==0.7.5
pure-eval==0.2.2
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-html==0.10.0
requests-oauthlib==1.3.1
streamlit==1.8.1
wordcloud==1.8.1